Create MemoryCacheService.
