+++ coil-base/src/main/java/coil/size/Size.kt
@@ -53,4 +53,5 @@ inline fun PixelSize(@Px width: Int, @Px height: Int) = Size(width, height)
     message = "Migrate to 'coil.size.Size.ORIGINAL'.",
     replaceWith = ReplaceWith("Size.ORIGINAL", "coil.size.Size")
 )
-inline val OriginalSize get() = Size.ORIGINAL
+inline val OriginalSize: Size
+    @JvmName("OriginalSize") get() = Size.ORIGINAL
